LPS-75270 Move upgrades to the packages they belong
