+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/internal/upgrade/v1_2_0/UpgradeDDLRecordSetSettingsFieldValues.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.internal.upgrade.v1_1_0;
+package com.liferay.dynamic.data.lists.internal.upgrade.v1_2_0;
 
 import com.liferay.dynamic.data.lists.model.DDLRecordSetConstants;
 import com.liferay.portal.kernel.json.JSONArray;
