+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-service/src/main/resources/com/liferay/dynamic/data/lists/internal/upgrade/v1_1_0/dependencies/update.sql
